# openapi.yaml
openapi: 3.0.3
info:
  title: Agent Memory Broker API
  version: 0.1.0
servers:
  - url: http://localhost:7070
paths:
  /agents:
    get:
      summary: List agents
      responses:
        '200': { description: OK }
    post:
      summary: Register agent
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AgentCreate' }}}}
      responses: { '200': { description: OK } }
  /contexts:
    get:
      summary: List context pools
      responses: { '200': { description: OK } }
    post:
      summary: Create context pool
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ContextCreate' }}}}
      responses: { '200': { description: OK } }
  /contexts/{context_id}/memories:
    parameters:
      - name: context_id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: List or query memories
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses: { '200': { description: OK } }
    post:
      summary: Append memory item
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MemoryCreate' }}}}
      responses: { '200': { description: OK } }
  /contexts/{context_id}/repos:
    parameters:
      - name: context_id
        in: path
        required: true
        schema: { type: string }
    post:
      summary: Link repository reference to context
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RepoRef' }}}}
      responses: { '200': { description: OK } }
  /contexts/{context_id}/tasks:
    parameters:
      - name: context_id
        in: path
        required: true
        schema: { type: string }
    post:
      summary: Create task
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TaskCreate' }}}}
      responses: { '200': { description: OK } }
    get:
      summary: List tasks
      responses: { '200': { description: OK } }
  /contexts/{context_id}/tasks/{task_id}:
    parameters:
      - name: context_id
        in: path
        required: true
        schema: { type: string }
      - name: task_id
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Update task
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TaskUpdate' }}}}
      responses: { '200': { description: OK } }
  /handoffs:
    post:
      summary: Record handoff
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/HandoffCreate' }}}}
      responses: { '200': { description: OK } }
  /contexts/{context_id}/handoffs:
    parameters:
      - name: context_id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: List handoffs
      responses: { '200': { description: OK } }
  /orchestrations:
    post:
      summary: Create orchestration run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/OrchestrationCreate'
                - type: object
                  properties:
                    plan:
                      $ref: '#/components/schemas/PlanV1'
      responses: { '200': { description: OK } }
  /orchestrations/{run_id}:
    parameters:
      - name: run_id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get orchestration run
      responses: { '200': { description: OK } }
  /orchestrations/{run_id}/events:
    parameters:
      - name: run_id
        in: path
        required: true
        schema: { type: string }
      - name: agent_id
        in: query
        schema: { type: string }
      - name: type
        in: query
        schema: { type: string }
      - name: category
        in: query
        schema: { type: string }
      - name: tag
        in: query
        schema: { type: string }
      - name: limit
        in: query
        schema: { type: integer, minimum: 0 }
      - name: after
        in: query
        schema: { type: string }
    get:
      summary: List events for orchestration run
      responses: { '200': { description: OK } }
  /orchestrations/{run_id}/cancel:
    parameters:
      - name: run_id
        in: path
        required: true
        schema: { type: string }
    post:
      summary: Cancel an orchestration run
      responses: { '200': { description: OK } }
  /contexts/{context_id}/events:
    parameters:
      - name: context_id
        in: path
        required: true
        schema: { type: string }
      - name: agent_id
        in: query
        schema: { type: string }
      - name: type
        in: query
        schema: { type: string }
      - name: category
        in: query
        schema: { type: string }
      - name: tag
        in: query
        schema: { type: string }
      - name: limit
        in: query
        schema: { type: integer, minimum: 0 }
      - name: after
        in: query
        schema: { type: string }
    get:
      summary: List events for context
      responses: { '200': { description: OK } }
components:
  schemas:
    AgentCreate:
      type: object
      required: [ name ]
      properties:
        name: { type: string }
        kind: { type: string, enum: [human, agent], default: agent }
        metadata: { type: object, additionalProperties: true }
    ContextCreate:
      type: object
      required: [ name ]
      properties:
        name: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
    MemoryCreate:
      type: object
      required: [ author, text ]
      properties:
        author: { type: string }
        text: { type: string }
        tags: { type: array, items: { type: string } }
        refs: { type: array, items: { $ref: '#/components/schemas/Ref' } }
    Ref:
      type: object
      properties:
        type: { type: string, enum: [file, url, repo, issue, pr] }
        value: { type: string }
        meta: { type: object, additionalProperties: true }
    RepoRef:
      type: object
      required: [ provider, owner, name ]
      properties:
        provider: { type: string, enum: [github] }
        owner: { type: string }
        name: { type: string }
        branch: { type: string }
    TaskCreate:
      type: object
      required: [ title ]
      properties:
        title: { type: string }
        description: { type: string }
        assignee: { type: string }
        status: { type: string, enum: [todo, in_progress, blocked, done], default: todo }
        tags: { type: array, items: { type: string } }
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        assignee: { type: string }
        status: { type: string, enum: [todo, in_progress, blocked, done] }
        tags: { type: array, items: { type: string } }
    HandoffCreate:
      type: object
      required: [ from, to, context, task, summary ]
      properties:
        from: { type: string }
        to: { type: string }
        context: { type: string }
        task: { type: string }
        summary: { type: string }
        next_steps: { type: string }
    OrchestrationPolicy:
      type: object
      properties:
        escalate_on_impasse: { type: boolean, default: true }
        max_agents: { type: integer }
        max_depth: { type: integer }
        extra: { type: object, additionalProperties: true }
    OrchestrationCreate:
      type: object
      required: [ context_id, prompt ]
      properties:
        context_id: { type: string }
        prompt: { type: string }
        created_by: { type: string, default: user }
        policy: { $ref: '#/components/schemas/OrchestrationPolicy' }
    PlanV1:
      type: object
      required: [ version ]
      properties:
        version: { type: string, enum: [ '1' ] }
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AgentSpec'
        actions:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/CreateAgentAction'
              - $ref: '#/components/schemas/MessageAction'
              - $ref: '#/components/schemas/TaskCreateAction'
              - $ref: '#/components/schemas/HandoffAction'
              - $ref: '#/components/schemas/SubprocessRunAction'
    AgentSpec:
      type: object
      required: [ name ]
      additionalProperties: false
      properties:
        name: { type: string }
        role: { type: string }
        kind: { type: string, enum: [ human, agent ], default: agent }
        id: { type: string }
        metadata: { type: object, additionalProperties: true }
    CreateAgentAction:
      type: object
      additionalProperties: false
      required: [ type, spec ]
      properties:
        type: { type: string, enum: [ create_agent ] }
        spec: { $ref: '#/components/schemas/AgentSpec' }
    MessageAction:
      type: object
      additionalProperties: false
      required: [ type, agent_id, content ]
      properties:
        type: { type: string, enum: [ message ] }
        agent_id: { type: string }
        content: { type: string }
    TaskCreateAction:
      type: object
      additionalProperties: false
      required: [ type, context_id, payload ]
      properties:
        type: { type: string, enum: [ 'task.create' ] }
        context_id: { type: string }
        payload: { $ref: '#/components/schemas/TaskCreate' }
    HandoffAction:
      type: object
      additionalProperties: false
      required: [ type, payload ]
      properties:
        type: { type: string, enum: [ handoff ] }
        payload: { $ref: '#/components/schemas/HandoffCreate' }
    SubprocessRunAction:
      type: object
      additionalProperties: false
      required: [ type, command ]
      properties:
        type: { type: string, enum: [ 'subprocess.run' ] }
        command: { type: array, items: { type: string } }
        env: { type: object, additionalProperties: { type: string } }
        cwd: { type: string }
    OrchestrationRun:
      type: object
      required: [ id, context_id, prompt, created_by, status, created_at ]
      properties:
        id: { type: string }
        context_id: { type: string }
        prompt: { type: string }
        created_by: { type: string }
        status: { type: string, enum: [pending, running, completed, failed] }
        policy: { $ref: '#/components/schemas/OrchestrationPolicy' }
        created_at: { type: string, format: date-time }
    Event:
      type: object
      required: [ id, context_id, category, type, actor, created_at ]
      properties:
        id: { type: string }
        context_id: { type: string }
        run_id: { type: string }
        category: { type: string, enum: [user, system, broker, agent, repo, task, handoff, memory, orchestration, plan] }
        type: { type: string }
        actor: { type: string, enum: [user, broker, agent] }
        message: { type: string }
        tags: { type: array, items: { type: string } }
        agent_id: { type: string }
        task_id: { type: string }
        handoff_id: { type: string }
        repo: { $ref: '#/components/schemas/RepoRef' }
        data: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }